message_editing: Fix topic edit propagate option cut-off.
In the topic changing UI, the 'Change later messages to this topic'
and other options in the drop down were cut off
when selected in some languages.
This commit makes the selection box width adjust
appropriately according to the length of the text.
It also moves it down to a new line for all users.