message-editing: Show options by default while editing topic only.
When trying to move message(s) by editing the topic
(after content editing has been disabled), the options
regarding the movement of messages related to that topic
show up only when at least some change is made to the topic
(not by default). This can hide the existence of 'movement of
messages' feature from the user. This commit shows these
options by default.